move _WorkerIndexer into libensemble.utils, also use within PersistentSupport
